move components around for clarity's sake
